Update test.yml
Removed "on push" trigger, as dependabot pushes was failing (as designed since such runs are running with read_only access and the CodeQL analysis step attempts to upload results to the repo). Needs a better long term fix if the maintenance will be continued, but this repo may likely be retired soon.